% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_construct.R
\name{custom_callout}
\alias{custom_callout}
\title{Insert custom callout left fence}
\usage{
custom_callout(
  title,
  content,
  iconify,
  collapse = TRUE,
  colorstyle = NULL,
  style = NULL,
  titleStyle
)
}
\arguments{
\item{title}{Title.}

\item{content}{Content in Markdown.}

\item{iconify}{Iconify icon to display in title.}

\item{collapse}{Default to TRUE.}

\item{colorstyle}{API to override font/border and background color, input should be string, which will pass to Quarto fence container. Use \code{var(--clcolor)} for font/border color and \code{var(--clgb)} for background color.}

\item{style}{Additional css styling, will be pass to \code{style=""}.}
}
\description{
Currently a new custom callout is not supported. This workaround uses the least used callout .callout-caution to mimic a custom callout
see: https://stackoverflow.com/questions/74647399/define-a-new-callout-in-quarto
}
\examples{
content <- ""
custom_callout("New Disclaimer", content, collapse = FALSE)
}
