% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_construct.R
\name{forwardMsg}
\alias{forwardMsg}
\title{Print Forward Message.}
\usage{
forwardMsg(msg, style)
}
\arguments{
\item{msg}{Message to print. If not proviede, use description in YAML header.}

\item{style}{Styling CSS to be passed to \verb{style=}.}
}
\description{
Print Forward Message.
}
