% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_construct.R
\name{header_image}
\alias{header_image}
\title{If header image is enable, the style has to be fixed. This could somehow not be done by calling external css file. It has to be specified in the same .qmd body.}
\usage{
header_image(height, y_transform)
}
\description{
Use \code{header_image_yaml()} to get the component for yaml header.
}
