% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_construct.R
\name{r_pkg_info}
\alias{r_pkg_info}
\title{Print out R session info and package citations}
\usage{
r_pkg_info(pkgs)
}
\arguments{
\item{pkgs}{A vector of package names.}
}
\description{
Print out R session info and package citations
}
